requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
colorama==0.4.6
requests==2.31.0
